+++ src/auth/dto/auth.dto.ts
@@ -1,0 +1,14 @@
+import { IsString } from 'class-validator';
+
+export class UserId {
+  @IsString()
+  userId: string;
+}
+
+export class VerifyCodeDto {
+  @IsString()
+  userId: string;
+
+  @IsString()
+  code: string;
+}
