+++ src/auth/repository/auth.type.ts
@@ -1,0 +1,3 @@
+export interface UserEmail {
+  email: string;
+}
